Refactor for getUser custom hook fetching data from dashboard to Layout
